Last commit, the game is done
You know, I must say, I'm glad it's over. But I'm also unbelievably happy that it happened. I am positively tired and exhausted and ready to move on to other projects, but this was still nice. I still...care about this game. I don't really know if it sucks because I haven't played it enough to know. But I hope it doesn't. I like this game, I like the idea and I like the people I worked with on it. So once again

Glad it's over, But glad it happened.

Anyways, I supposed it's showtime. Time to release this. Our work, our love our child to the public. Or to the very few people who'll see this on itch.io. Anyways, at least for now, this is goodbye from Xara Neale Shaw.